Test: split seq and map engine tests to flow and block
