+++ test/CMakeLists.txt
@@ -92,9 +92,10 @@ endfunction()
 
 
 ryml_add_test(engine_1_doc)
-ryml_add_test(engine_2_seq)
-ryml_add_test(engine_3_map_1)
-ryml_add_test(engine_3_map_2)
+ryml_add_test(engine_2_seq_block)
+ryml_add_test(engine_2_seq_flow)
+ryml_add_test(engine_3_map_block)
+ryml_add_test(engine_3_map_flow)
 ryml_add_test(engine_4_anchor)
 ryml_add_test(engine_5_tag)
 ryml_add_test(engine_6_qmrk_1)
@@ -163,9 +164,10 @@ endif()
 # MSVC bigobj
 if(MSVC)
     target_compile_options(ryml-test-engine_1_doc            PRIVATE /bigobj)
-    target_compile_options(ryml-test-engine_2_seq            PRIVATE /bigobj)
-    target_compile_options(ryml-test-engine_3_map_1          PRIVATE /bigobj)
-    target_compile_options(ryml-test-engine_3_map_2          PRIVATE /bigobj)
+    target_compile_options(ryml-test-engine_2_seq_block      PRIVATE /bigobj)
+    target_compile_options(ryml-test-engine_2_seq_flow       PRIVATE /bigobj)
+    target_compile_options(ryml-test-engine_3_map_block      PRIVATE /bigobj)
+    target_compile_options(ryml-test-engine_3_map_flow       PRIVATE /bigobj)
     target_compile_options(ryml-test-engine_4_anchor         PRIVATE /bigobj)
     target_compile_options(ryml-test-engine_5_tag            PRIVATE /bigobj)
     target_compile_options(ryml-test-engine_6_qmrk_1         PRIVATE /bigobj)
